requirements.txt
status
Django==1.9
django-filter==0.13.0
djangorestframework==3.3.3
Markdown==2.6.6
Pillow==3.1.0
